requirements.txt
asgiref==3.5.2
deprecation==2.1.0
Django==4.1.2
docx2pdf==0.1.8
fpdf==1.7.2
gunicorn==20.1.0
img2pdf==0.4.4
lxml==4.9.1
packaging==21.3
pikepdf==6.2.1
Pillow==9.2.0
pyparsing==3.0.9
PyPDF2==2.11.1
sqlparse==0.4.3
tzdata==2022.5
